backend/requirements.txt
anaconda==0.0.1.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
binaryornot==0.4.4
bs4==0.0.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
Django==5.1.7
django-allauth==65.3.1
django-appconf==1.0.6
django-browser-reload==1.18.0
django-ckeditor==6.7.2
django-compressor==4.5.1
django-cors-headers==4.3.1
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
exchange_calendars==4.10
executing==2.2.0
Faker==37.1.0
fonttools==4.56.0
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
MetaTrader5==5.0.4874
msgpack==1.1.0
multitasking==0.0.11
mysql==0.0.3
mysqlclient==2.2.7
narwhals==1.32.0
nest-asyncio==1.6.0
nselib==1.8
numpy==2.1.3
oauthlib==3.2.2
openai==1.63.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas_market_calendars==4.6.1
pandasgui==0.2.14
parso==0.8.4
pdfkit==1.0.0
peewee==3.17.9
phonenumbers==8.13.54
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyluach==2.2.0
pynput==1.8.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
qtstylish==0.1.5
rcssmin==1.1.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rjsmin==1.2.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
service-identity==24.2.0
setuptools==75.9.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wordcloud==1.9.4
wsproto==1.2.0
yfinance==0.2.55
zope.interface==7.2
gunicorn==21.2.0
whitenoise==6.6.0
